Improved performance and fixed disappearing numbers
This commit fixes two things. First, I did not properly test the piece of code responsible for skipping updating cards when an update was not needed. So that's a huge performance win.

Secondly, I fixed an issue where updating a card would remove the counter on another card. It was caused by resetting the data-mutated property on the wrong element. Difficult to debug, easy to fix once found.

Fixed #5